requirements.txt
tqdm
opencv-python
ftfy
regex
matplotlib
ttach
scipy
ipywidgets
timm
pandas
pillow